move dnd into ds primitives
